+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64/FitCloudKit.framework/Headers/FitCloudKit.h
@@ -10,7 +10,7 @@
 //          FitCloudPro 智能手表 iOS 框架，封装了与手表设备通信等核心功能。
 //
 //  构建版本：
-//      pcjbird    2026-06-09  Version:1.3.2-beta.72 Build:20260609001
+//      pcjbird    2026-06-10  Version:1.3.2-beta.73 Build:20260610001
 
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>